fix: refactor IVariableTreeNode
