+++ packages/helpers/src/types/advanced.ts
@@ -34,6 +34,52 @@ export type OptionType = {
 };
 
 /**
+ * 变量树节点类型
+ */
+export interface IVariableTreeNode {
+  /**
+   * 唯一标识符
+   */
+  key: string;
+  /**
+   * 标题
+   */
+  title?: string;
+  /**
+   * 辅助提示信息
+   */
+  help?: string;
+  /**
+   * 是否可选中
+   */
+  selectable?: boolean;
+  /**
+   * 展示删除按钮
+   */
+  showRemoveButton?: boolean;
+  /**
+   * 展示添加按钮
+   */
+  showAddButton?: boolean;
+  /**
+   * 展示查看按钮
+   */
+  showViewButton?: boolean;
+  /**
+   * 结点类型，用来展示图标
+   */
+  type?: 'function' | 'property';
+  /**
+   * 定义的原始值
+   */
+  raw?: any;
+  /**
+   * 子结点
+   */
+  children?: IVariableTreeNode[];
+}
+
+/**
  * 服务函数类型
  */
 export type ServiceFunctionType = (...args: any[]) => Promise<any>;
